Extract gesture handler to own file
